+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irBlockBody.kt
@@ -1,0 +1,12 @@
+package com.javiersc.kotlin.compiler.extensions.ir
+
+import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
+import org.jetbrains.kotlin.ir.UNDEFINED_OFFSET
+import org.jetbrains.kotlin.ir.declarations.createBlockBody
+import org.jetbrains.kotlin.ir.expressions.IrBlockBody
+
+public fun IrPluginContext.createIrBlockBody(
+    startOffset: Int = UNDEFINED_OFFSET,
+    endOffset: Int = UNDEFINED_OFFSET,
+    initializer: IrBlockBody.() -> Unit,
+): IrBlockBody = irFactory.createBlockBody(startOffset, endOffset, initializer)