+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irFunctionExpression.kt
@@ -1,0 +1,23 @@
+package com.javiersc.kotlin.compiler.extensions.ir
+
+import org.jetbrains.kotlin.ir.UNDEFINED_OFFSET
+import org.jetbrains.kotlin.ir.declarations.IrSimpleFunction
+import org.jetbrains.kotlin.ir.expressions.IrFunctionExpression
+import org.jetbrains.kotlin.ir.expressions.IrStatementOrigin
+import org.jetbrains.kotlin.ir.expressions.impl.IrFunctionExpressionImpl
+import org.jetbrains.kotlin.ir.types.IrType
+
+public fun createIrFunctionExpression(
+    startOffset: Int = UNDEFINED_OFFSET,
+    endOffset: Int = UNDEFINED_OFFSET,
+    type: IrType,
+    function: IrSimpleFunction,
+    origin: IrStatementOrigin
+): IrFunctionExpression =
+    IrFunctionExpressionImpl(
+        startOffset = startOffset,
+        endOffset = endOffset,
+        type = type,
+        function = function,
+        origin = origin,
+    )