+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/ir/irValueParameter.kt
@@ -1,5 +1,14 @@
 package com.javiersc.kotlin.compiler.extensions.ir
 
-//fun IrCall.mapValueParameters(
-//    map: (IrValueParameter) -> IrValueParameter,
-//): IrCall = copyValueArguments(valueArguments.map(map))
+import org.jetbrains.kotlin.ir.UNDEFINED_OFFSET
+import org.jetbrains.kotlin.ir.declarations.IrValueParameter
+import org.jetbrains.kotlin.ir.expressions.IrGetValue
+import org.jetbrains.kotlin.ir.expressions.impl.IrGetValueImpl
+
+public fun IrValueParameter.toIrGetValue(): IrGetValue =
+    IrGetValueImpl(
+        startOffset = UNDEFINED_OFFSET,
+        endOffset = UNDEFINED_OFFSET,
+        type = symbol.owner.type,
+        symbol = symbol
+    )